Refactor alignment pipeline into private modules
