+++ src/tomojax/align/_setup_stage.py
@@ -1,0 +1,304 @@
+from __future__ import annotations
+
+import math
+from typing import Iterable
+
+import jax.numpy as jnp
+import numpy as np
+
+from ..core.geometry.axis import RotationAxisGeometry
+from ..core.geometry.base import Detector, Geometry, Grid
+from ..core.geometry.lamino import LaminographyGeometry
+from ..core.geometry.parallel import ParallelGeometry
+from ..recon.fista_tv import FistaConfig, fista_tv
+from ._observer import OuterStat
+from .diagnostics import validate_active_gauge_policy
+from .dof_specs import ActiveParameterView
+from .fold_recon import FoldReconstructionConfig, reconstruct_train_fold_nograd
+from .folds import FoldSpec
+from .geometry_applier import BaseGeometryArrays, apply_setup_to_detector_grid, setup_axis_unit
+from ._loss_adapters import build_loss_adapter
+from .optimizers import ValidationLmConfig, run_active_validation_lm
+from .schedules import ResolvedAlignmentStage
+from .state import AlignmentState, PoseState, SetupGeometryState
+from .validation_residuals import accumulate_validation_normals, score_validation_fixed_volume
+
+
+def _geometry_with_setup_state(
+    geometry: Geometry,
+    grid: Grid,
+    detector: Detector,
+    setup: SetupGeometryState,
+) -> Geometry:
+    thetas = np.asarray(getattr(geometry, "thetas_deg"), dtype=np.float32)
+    axis = tuple(float(v) for v in np.asarray(setup_axis_unit(setup), dtype=np.float32))
+    axis_active = (
+        abs(float(setup.axis_rot_x_rad)) > 1e-7
+        or abs(float(setup.axis_rot_y_rad)) > 1e-7
+        or isinstance(geometry, RotationAxisGeometry)
+    )
+    if axis_active:
+        return RotationAxisGeometry(
+            grid=grid,
+            detector=detector,
+            thetas_deg=thetas,
+            axis_unit_lab=axis,
+        )
+    if isinstance(geometry, LaminographyGeometry):
+        return LaminographyGeometry(
+            grid=grid,
+            detector=detector,
+            thetas_deg=thetas,
+            tilt_deg=float(geometry.tilt_deg),
+            tilt_about=str(geometry.tilt_about),
+        )
+    return ParallelGeometry(grid=grid, detector=detector, thetas_deg=thetas)
+
+
+def _geometry_calibration_payload(
+    state: AlignmentState,
+    active_geometry_dofs: Iterable[str],
+) -> dict[str, object]:
+    return state.to_calibration_state(active_dofs=active_geometry_dofs).to_dict()
+
+
+def _optimize_setup_geometry_bilevel_for_level(
+    *,
+    geometry: Geometry,
+    grid: Grid,
+    detector: Detector,
+    projections: jnp.ndarray,
+    init_x: jnp.ndarray | None,
+    init_params5: jnp.ndarray | None,
+    state: AlignmentState,
+    active_geometry_dofs: Iterable[str],
+    factor: int,
+    cfg: object,
+    loss_spec,
+    loss_name: str,
+    schedule_name: str | None = None,
+    stage: ResolvedAlignmentStage | None = None,
+) -> tuple[jnp.ndarray, AlignmentState, list[OuterStat]]:
+    base = BaseGeometryArrays.from_geometry(geometry, detector, level_factor=int(factor))
+    active_view = ActiveParameterView.from_dofs(active_geometry_dofs, geometry=geometry)
+    alignment_state = state.replace(
+        setup=state.setup.replace(nominal_axis_unit=base.nominal_axis_unit),
+        pose=PoseState(
+            jnp.zeros((int(projections.shape[0]), 5), dtype=jnp.float32)
+            if init_params5 is None
+            else jnp.asarray(init_params5, dtype=jnp.float32)
+        ),
+        volume=init_x,
+    )
+    n_views = int(projections.shape[0])
+    n_folds = min(2, n_views)
+    folds = FoldSpec(n_folds=n_folds).build(n_views)
+    loss_adapter = build_loss_adapter(loss_spec, projections)
+    if not bool(loss_adapter.supports_setup_validation_lm):
+        stage_name = schedule_name or (stage.name if stage is not None else "setup")
+        raise ValueError(
+            "Setup validation-LM requires a setup-compatible weighted least-squares loss; "
+            f"level {int(factor)} stage {stage_name!r} resolved loss {loss_name!r}"
+        )
+    fold_recon_cfg = FoldReconstructionConfig(
+        iters=max(1, int(cfg.recon_iters)),
+        lambda_tv=float(cfg.lambda_tv),
+        regulariser=str(cfg.regulariser),
+        huber_delta=float(cfg.huber_delta),
+        tv_prox_iters=int(cfg.tv_prox_iters),
+        L=cfg.recon_L,
+        positivity=bool(cfg.recon_positivity),
+        views_per_batch=max(1, int(cfg.views_per_batch)),
+        projector_unroll=int(cfg.projector_unroll),
+        checkpoint_projector=bool(cfg.checkpoint_projector),
+        gather_dtype=str(cfg.gather_dtype),
+    )
+
+    setup_state = alignment_state
+    setup_stats: list[OuterStat] = []
+    last_loss = math.inf
+    outer_limit = max(1, int(stage.maxiter if stage is not None else cfg.outer_iters))
+    for outer_idx in range(1, outer_limit + 1):
+        z_current = active_view.pack(setup_state)
+        total_loss = jnp.asarray(0.0, dtype=jnp.float32)
+        total_grad = jnp.zeros_like(z_current)
+        total_hess = jnp.zeros((int(z_current.size), int(z_current.size)), dtype=jnp.float32)
+        residual_count = 0
+        fold_cache: list[tuple[int, jnp.ndarray, jnp.ndarray, jnp.ndarray, dict[str, object]]] = []
+        for fold in range(folds.n_folds):
+            train_idx = folds.train_idx[fold]
+            train_mask = folds.train_mask[fold]
+            val_idx = folds.val_idx[fold]
+            val_mask = folds.val_mask[fold]
+            fold_volume, fold_recon_info = reconstruct_train_fold_nograd(
+                geometry=geometry,
+                grid=grid,
+                detector=detector,
+                projections=projections,
+                state=setup_state,
+                train_idx=train_idx,
+                train_mask=train_mask,
+                init_x=init_x,
+                level_factor=int(factor),
+                cfg=fold_recon_cfg,
+            )
+            normals = accumulate_validation_normals(
+                frozen_state=setup_state,
+                active_view=active_view,
+                z=z_current,
+                base=base,
+                grid=grid,
+                detector=detector,
+                projections=projections,
+                loss_adapter=loss_adapter,
+                fold_volume=fold_volume,
+                val_idx=val_idx,
+                val_mask=val_mask,
+                views_per_batch=max(1, int(cfg.views_per_batch)),
+                projector_unroll=int(cfg.projector_unroll),
+                checkpoint_projector=bool(cfg.checkpoint_projector),
+                gather_dtype=str(cfg.gather_dtype),
+            )
+            total_loss = total_loss + normals.loss
+            total_grad = total_grad + normals.grad
+            total_hess = total_hess + normals.hess
+            residual_count += int(normals.residual_count)
+            fold_cache.append((fold, fold_volume, val_idx, val_mask, fold_recon_info))
+
+        def score_candidate(z_candidate: jnp.ndarray) -> float:
+            score = jnp.asarray(0.0, dtype=jnp.float32)
+            for _fold, fold_volume, val_idx, val_mask, _fold_info in fold_cache:
+                score = score + score_validation_fixed_volume(
+                    frozen_state=setup_state,
+                    active_view=active_view,
+                    z=z_candidate,
+                    base=base,
+                    grid=grid,
+                    detector=detector,
+                    projections=projections,
+                    loss_adapter=loss_adapter,
+                    fold_volume=fold_volume,
+                    val_idx=val_idx,
+                    val_mask=val_mask,
+                    views_per_batch=max(1, int(cfg.views_per_batch)),
+                    projector_unroll=int(cfg.projector_unroll),
+                    checkpoint_projector=bool(cfg.checkpoint_projector),
+                    gather_dtype=str(cfg.gather_dtype),
+                )
+            return float(score)
+
+        opt_result = run_active_validation_lm(
+            state=setup_state,
+            view=active_view,
+            loss=float(total_loss),
+            grad=total_grad,
+            hess=total_hess,
+            score_fn=score_candidate,
+            bounds=cfg.bounds,
+            cfg=ValidationLmConfig(damping=max(float(cfg.gn_damping), 1e-6)),
+        )
+        setup_state = opt_result.state
+        last_loss = float(opt_result.loss)
+        stat = dict(opt_result.stats)
+        stat.update(
+            {
+                "geometry_block": "setup_validation_lm",
+                "geometry_active_dofs": ",".join(active_view.dofs),
+                "geometry_objective": "bilevel_cv",
+                "geometry_optimizer": "validation_lm",
+                "geometry_loss_kind": loss_name,
+                "geometry_loss_before": float(total_loss),
+                "geometry_loss_after": float(opt_result.loss),
+                "geometry_accepted": bool(opt_result.accepted),
+                "geometry_step_norm": float(stat.get("step_norm_whitened", 0.0) or 0.0),
+                "geometry_gradient_norm": float(stat.get("grad_norm_whitened", 0.0) or 0.0),
+                "geometry_max_step": 1.0,
+                "geometry_status": "converged" if opt_result.accepted else "underconverged",
+                "geometry_outer_idx": int(outer_idx),
+                "schedule_name": schedule_name,
+                "schedule_stage_index": (
+                    int(stage.index) if stage is not None else None
+                ),
+                "schedule_stage_name": stage.name if stage is not None else None,
+                "schedule_stage_active_dofs": (
+                    ",".join(stage.active_dofs) if stage is not None else ",".join(active_view.dofs)
+                ),
+                "gauge_policy": stage.gauge_policy if stage is not None else cfg.gauge_policy,
+                "gauge_status": (
+                    stage.gauge_decision.status
+                    if stage is not None
+                    else validate_active_gauge_policy(
+                        active_view.dofs,
+                        policy=cfg.gauge_policy,
+                        priors=cfg.gauge_priors,
+                    ).status
+                ),
+                "gauge_decision": (
+                    stage.gauge_decision.to_dict()
+                    if stage is not None
+                    else validate_active_gauge_policy(
+                        active_view.dofs,
+                        policy=cfg.gauge_policy,
+                        priors=cfg.gauge_priors,
+                    ).to_dict()
+                ),
+                "objective_kind": "bilevel_cv",
+                "objective_provenance": {
+                    "outer_loss_source": "AlignmentLossSpec",
+                    "outer_loss_kind": str(loss_name),
+                    "inner_data_term": "l2_projection",
+                    "inner_regulariser": str(fold_recon_cfg.regulariser),
+                    "validation_split": "interleaved_kfold",
+                    "differentiation_mode": "none",
+                    "initialization_policy": "current_level_volume" if init_x is not None else "zeros",
+                },
+                "optimizer_kind": "validation_lm",
+                "outer_loss_kind": str(loss_name),
+                "recon_sensitivity": "stopped",
+                "train_reconstruction_gradient": False,
+                "views_per_batch": max(1, int(cfg.views_per_batch)),
+                "n_folds": int(folds.n_folds),
+                "fold_eval_mode": "stopped_train_recon_validation_lm",
+                "folds_used": ",".join(str(item[0]) for item in fold_cache),
+                "num_train_reconstructions": int(len(fold_cache)),
+                "validation_residual_count": int(residual_count),
+                "recon_projection_chunked": True,
+                "validation_projection_chunked": True,
+                "active_gradient_mode": "validation_residual_jvp",
+            }
+        )
+        setup_stats.append(stat)
+        if bool(cfg.early_stop) and outer_idx > 1:
+            prev = float(setup_stats[-2].get("geometry_loss_after", math.inf))
+            impr = (prev - last_loss) / max(abs(prev), 1e-6)
+            if impr < float(cfg.early_stop_rel_impr):
+                break
+
+    geom = _geometry_with_setup_state(geometry, grid, detector, setup_state.setup)
+    det_grid = apply_setup_to_detector_grid(
+        detector,
+        setup_state.setup,
+        level_factor=int(factor),
+    )
+    x_next, _ = fista_tv(
+        geom,
+        grid,
+        detector,
+        projections,
+        init_x=init_x,
+        config=FistaConfig(
+            iters=max(1, int(cfg.recon_iters)),
+            lambda_tv=float(cfg.lambda_tv),
+            regulariser=cfg.regulariser,
+            huber_delta=float(cfg.huber_delta),
+            tv_prox_iters=int(cfg.tv_prox_iters),
+            L=cfg.recon_L,
+            views_per_batch=max(1, int(cfg.views_per_batch)),
+            projector_unroll=int(cfg.projector_unroll),
+            checkpoint_projector=bool(cfg.checkpoint_projector),
+            gather_dtype=str(cfg.gather_dtype),
+            positivity=bool(cfg.recon_positivity),
+        ),
+        det_grid=det_grid,
+    )
+    return x_next, setup_state.replace(volume=x_next), setup_stats
